Phillip Feist's TechU Coding Challenge Submission
Hello,

Here are my submissions for the TechU Challenge.  The ninja bookshelf
puzzle has had me baffled.  I was able to get the program to print my
book names vertically as requested, but when it reaches the end of the
shortest title the program quits and throws an error.  I tried
recursion, using the class Integer instead of int for i, converting my
char output into String output, and a whole variety of for loops with
if/else stuctures, but to no avail.  When we meet I would like some
guidance on how this can be overcome, just to satisfy my curiosity.  I
will attempt to solve at least one bonus challenge and submit before
Monday, but in the mean time have a good weekend and I hope to hear
from you soon.

Thanks,

Phil Feist